Fix normal -> sealable namespace

+++ internal/vault/namespace_store.go
@@ -25,7 +25,6 @@ import (
 	"github.com/openbao/openbao/v2/internal/helper/namespace"
 	"github.com/openbao/openbao/v2/internal/vault/barrier"
 	"github.com/openbao/openbao/v2/internal/vault/policy"
-	vaultseal "github.com/openbao/openbao/v2/internal/vault/seal"
 )
 
 // Namespace id length; upstream uses 5 characters so we use one more to
@@ -349,6 +348,7 @@ func (ns *NamespaceStore) Invalidate(ctx context.Context, parentUUID, childUUID 
 		}
 	}
 
+	b = ns.core.sealManager.NamespaceBarrierByLongestPrefix(child.Path)
 	ns.core.router.SetNamespaceStorageView(&child, b)
 
 	if child.ManuallySealed {
@@ -1937,59 +1937,31 @@ func (j *namespaceCreationFailureJob) OnFailure(err error) {
 }
 
 type namespaceBarrierMigrationJob struct {
-	store      *NamespaceStore
-	core       *Core
-	sealConfig *SealConfig
-	parent     *namespace.Namespace
-	target     *namespace.Namespace
+	store         *NamespaceStore
+	core          *Core
+	seal          Seal
+	sealConfig    *SealConfig
+	oldBarrier    barrier.SecurityBarrier
+	newBarrier    barrier.SecurityBarrier
+	parentBarrier barrier.SecurityBarrier
+	target        *namespace.Namespace
 }
 
-func (ns *NamespaceStore) newNamespaceBarrierMigrationJob(parent *namespace.Namespace, target *namespace.Namespace, sealConfig *SealConfig) fairshare.Job {
+func (ns *NamespaceStore) newNamespaceBarrierMigrationJob(parentBarrier, oldBarrier, newBarrier barrier.SecurityBarrier, target *namespace.Namespace, seal Seal, sealConfig *SealConfig) fairshare.Job {
 	return &namespaceBarrierMigrationJob{
-		store:      ns,
-		core:       ns.core,
-		sealConfig: sealConfig,
-		parent:     parent,
-		target:     target,
+		store:         ns,
+		core:          ns.core,
+		seal:          seal,
+		sealConfig:    sealConfig,
+		target:        target,
+		oldBarrier:    oldBarrier,
+		newBarrier:    newBarrier,
+		parentBarrier: parentBarrier,
 	}
 }
 
 func (j *namespaceBarrierMigrationJob) Execute() error {
-	ctx := namespace.ContextWithNamespace(j.store.asyncJobContext, j.target)
-
-	if err := j.store.taintNamespace(ctx, j.parent, j.target); err != nil {
-		return err
-	}
-	defer j.store.untaintNamespace(ctx, j.parent, j.target)
-
-	parentBarrier := j.core.sealManager.NamespaceBarrierByLongestPrefix(j.parent.Path)
-	oldBarrier := j.core.sealManager.NamespaceBarrierByLongestPrefix(j.target.Path)
-
-	var newBarrier barrier.SecurityBarrier
-	if j.sealConfig == nil {
-		newBarrier = parentBarrier
-	}
-
-	if newBarrier == nil {
-		metaPrefix := NamespaceStoragePathPrefix(j.target)
-		defaultSeal := NewDefaultSeal(vaultseal.NewAccess(vaultseal.NewShamirWrapper()))
-		defaultSeal.SetCore(j.core)
-		defaultSeal.SetMetaPrefix(metaPrefix)
-
-		defaultSeal.SetConfigAccess(parentBarrier)
-
-		ctx := namespace.ContextWithNamespace(ctx, j.target)
-		if err := defaultSeal.Init(ctx); err != nil {
-			return err
-		}
-
-		newBarrier = barrier.NewAESGCMBarrier(j.core.physical, j.target)
-	}
-
-	if newBarrier == oldBarrier {
-		// Nothing to do
-		return nil
-	}
+	ctx := j.store.asyncJobContext
 
 	namespacesToMigrate, err := j.store.ListNamespaces(namespace.ContextWithNamespace(ctx, j.target), ListNamespaceOpts{
 		IncludeSealed: false,
@@ -2004,10 +1976,10 @@ func (j *namespaceBarrierMigrationJob) Execute() error {
 	namespacesToMigrate = append(namespacesToMigrate, j.target)
 
 	var errs error
-	err = logical.WithTransaction(ctx, newBarrier, func(s logical.Storage) (err error) {
+	err = logical.WithTransaction(ctx, j.newBarrier, func(s logical.Storage) (err error) {
 		for _, ns := range namespacesToMigrate {
 			nsPrefix := NamespaceScopedView(s, ns).Prefix()
-			keys, err := recurseListKeys(ctx, oldBarrier, nsPrefix)
+			keys, err := recurseListKeys(ctx, j.oldBarrier, nsPrefix)
 			if err != nil {
 				return err
 			}
@@ -2016,14 +1988,16 @@ func (j *namespaceBarrierMigrationJob) Execute() error {
 				if slices.ContainsFunc(knownNamespaceCoreEntriesToCleanup, func(k string) bool {
 					return strings.HasSuffix(key, "/"+k)
 				}) {
-					err := s.Delete(ctx, key)
-					if err != nil {
-						return err
+					if j.newBarrier == j.parentBarrier {
+						err := s.Delete(ctx, key)
+						if err != nil {
+							return err
+						}
 					}
 					continue
 				}
 
-				se, err := oldBarrier.Get(ctx, key)
+				se, err := j.oldBarrier.Get(ctx, key)
 				if err != nil {
 					return err
 				}
@@ -2035,25 +2009,27 @@ func (j *namespaceBarrierMigrationJob) Execute() error {
 			}
 		}
 
-		switch newBarrier {
-		case parentBarrier:
-			// in case of an error, this will be reverted by the namespace invalidation at the end
-			j.core.sealManager.RemoveNamespace(j.target)
-		default:
-			// in case of an error, this will be reverted by the namespace invalidation at the end
-			err := j.core.sealManager.SetSeal(ctx, j.sealConfig, j.target, SetSealOptions{
-				WriteToStorage: true,
-				AllowOverride:  true,
-			})
-			if err != nil {
-				return err
-			}
-		}
-
 		return nil
 	})
 	if err != nil {
 		errs = errors.Join(errs, err)
+	}
+
+	switch j.newBarrier {
+	case j.parentBarrier:
+		// in case of an error, this will be reverted by the namespace invalidation at the end
+		j.core.sealManager.RemoveNamespace(j.target)
+	default:
+		// in case of an error, this will be reverted by the namespace invalidation at the end
+		err = j.core.sealManager.SetSeal(ctx, j.sealConfig, j.target, SetSealOptions{
+			WriteToStorage: true,
+			AllowOverride:  true,
+			Seal:           j.seal,
+			Barrier:        j.newBarrier,
+		})
+		if err != nil {
+			return err
+		}
 	}
 
 	slices.Reverse(namespacesToMigrate)
@@ -2082,5 +2058,5 @@ func (j *namespaceBarrierMigrationJob) Execute() error {
 }
 
 func (j *namespaceBarrierMigrationJob) OnFailure(err error) {
-	j.store.logger.Error("failed to handle namespace deletion following failed creation; job may be retried via deletion of tainted namespace", "namespace", j.target.Path, "ns_uuid", j.target.UUID, "error", err.Error())
+	j.store.logger.Error("failed to handle namespace barrier migration", "namespace", j.target.Path, "ns_uuid", j.target.UUID, "error", err.Error())
 }
